refactor : separate layout folder into Desktop and Mobile & move existiong files to Desktop folder
